refactor(types): extract User to models
