+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/models/index.d.ts
@@ -1,0 +1,1 @@
+export * from "./user";
