+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/models/user.d.ts
@@ -1,0 +1,8 @@
+import { Address } from "../types";
+
+export interface User {
+  name: string;
+  age: number;
+  email: string;
+  address: Address;
+}
