+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/types/index.d.ts
@@ -1,11 +1,4 @@
-interface Address {
+export interface Address {
   street: string;
   no: number;
-}
-
-export interface User {
-  name: string;
-  age: number;
-  email: string;
-  address: Address;
 }
